v1.1b now checks for dropbox existence

+++ sauce/templates/rump-update-and-execute.sh
@@ -1,9 +1,10 @@
 #!/bin/sh
 
-VER=1.1
+VER=1.1b
 
 # touchfile
-DROPBOX_FILE="<%= poweruser_home %>/Dropbox/tmp/sauce/cron-rump-<%= fqdn %>-v$VER.touch"
+mkdir -p "<%= poweruser_home %>/Dropbox/tmp/sauce/cron-rump/"
+DROPBOX_FILE="<%= poweruser_home %>/Dropbox/tmp/sauce/cron-rump/<%= fqdn %>-v$VER.touch"
 
 cd ~/git/puppet-rump && 
 	git pull origin master &&  
@@ -12,8 +13,8 @@ cd ~/git/puppet-rump &&
 	touch <%= basepath %>/cron-rump-last-update-v$VER.touch
 
 # If it exists I create a touch :)
-#if [ -f "<%= poweruser_home %>/Dropbox/tmp/sauce/" ]; then
+if [ -d "<%= poweruser_home %>/Dropbox/tmp/sauce/" ]; then
   rm "<%= poweruser_home %>/Dropbox/tmp/sauce/<%= hostname %>-cron-rump-last-update-v*.touch"
   touch "$DROPBOX_FILE"
   chmod <%=poweruser_name%> "$DROPBOX_FILE"
-#fi
+fi
